feat: Dedupe the protobuf binding work into `common-protos`, and improve incremental building.
