+++ rust/common-protos/README.md
@@ -1,0 +1,4 @@
+# common-protos
+
+Protobuf bindings.
+
