+++ rust/common-runtime/Cargo.toml
@@ -13,12 +13,12 @@ async-stream = { workspace = true }
 axum = { workspace = true, features = ["multipart", "macros"] }
 blake3 = { workspace = true }
 bytes = { workspace = true }
+common-protos = { workspace = true, features = ["runtime"] }
 common-tracing = { workspace = true }
 common-wit = { workspace = true }
 http = { workspace = true }
 hyper-util = { workspace = true }
 mime_guess = { workspace = true }
-prost = { workspace = true }
 rand = { workspace = true }
 redb = { workspace = true }
 rust-embed = { workspace = true }
@@ -40,6 +40,3 @@ wit-parser = { workspace = true }
 [dev-dependencies]
 common-builder = { workspace = true }
 common-test-fixtures = { workspace = true }
-
-[build-dependencies]
-tonic-build = { workspace = true }
