+++ rust/common-runtime/src/runtime.rs
@@ -1,10 +1,10 @@
-use std::collections::BTreeMap;
-
 use crate::{
     wasmtime::WasmtimeCompiler, CommonRuntimeError, InputOutput, ModuleDefinition, ModuleInstance,
     ModuleInstanceId, ModulePreparer, OutputShape, PreparedModule, ToModuleSources,
     ToWasmComponent, Value, ValueKind,
 };
+use common_protos::common;
+use std::collections::{BTreeMap, HashMap};
 
 /// An implementation of [InputOutput] that is suitable for use with a
 /// [Runtime].
@@ -134,5 +134,28 @@ impl Runtime {
         };
 
         instance.run(io).await
+    }
+}
+
+impl TryFrom<(HashMap<String, common::Value>, HashMap<String, i32>)> for RuntimeIo {
+    type Error = CommonRuntimeError;
+
+    fn try_from(
+        (input_proto, output_shape_proto): (HashMap<String, common::Value>, HashMap<String, i32>),
+    ) -> Result<Self, Self::Error> {
+        let mut input = BTreeMap::new();
+        for (key, value) in input_proto.into_iter() {
+            input.insert(key, Value::try_from(value)?);
+        }
+
+        let mut output_shape = BTreeMap::new();
+
+        for (key, value_kind) in output_shape_proto.into_iter() {
+            let value_kind = common::ValueKind::try_from(value_kind)
+                .map_err(|_| CommonRuntimeError::InvalidValue)?;
+            output_shape.insert(key, ValueKind::from(value_kind));
+        }
+
+        Ok(RuntimeIo::new(input, output_shape))
     }
 }
