+++ rust/common-runtime/src/value.rs
@@ -1,3 +1,6 @@
+use crate::CommonRuntimeError;
+use common_protos::common;
+
 /// An intrinsic value type within a Common Runtime
 #[derive(Clone, Debug)]
 pub enum Value {
@@ -35,4 +38,42 @@ pub enum ValueKind {
     Number,
     /// A slab of bytes
     Buffer,
+}
+
+impl TryFrom<common::Value> for Value {
+    type Error = CommonRuntimeError;
+
+    fn try_from(value: common::Value) -> Result<Self, Self::Error> {
+        let value = value.variant.ok_or(CommonRuntimeError::InvalidValue)?;
+        Ok(match value {
+            common::value::Variant::String(string) => Value::String(string),
+            common::value::Variant::Number(number) => Value::Number(number),
+            common::value::Variant::Boolean(boolean) => Value::Boolean(boolean),
+            common::value::Variant::Buffer(buffer) => Value::Buffer(buffer),
+        })
+    }
+}
+
+impl From<Value> for common::Value {
+    fn from(value: Value) -> Self {
+        common::Value {
+            variant: Some(match value {
+                Value::String(string) => common::value::Variant::String(string),
+                Value::Boolean(number) => common::value::Variant::Boolean(number),
+                Value::Number(boolean) => common::value::Variant::Number(boolean),
+                Value::Buffer(buffer) => common::value::Variant::Buffer(buffer),
+            }),
+        }
+    }
+}
+
+impl From<common::ValueKind> for ValueKind {
+    fn from(value_kind: common::ValueKind) -> Self {
+        match value_kind {
+            common::ValueKind::String => ValueKind::String,
+            common::ValueKind::Boolean => ValueKind::Boolean,
+            common::ValueKind::Number => ValueKind::Number,
+            common::ValueKind::Buffer => ValueKind::Buffer,
+        }
+    }
 }
